refactor: change name single sms zustand file
