+++ app/shared/state-management/single-sms.ts
@@ -18,14 +18,27 @@ export const useTag = create<Tags>(set => ({
     }),
   removeTag: (tag): void => set(prevState => ({ tags: prevState.tags.filter(item => item !== tag) }))
 }));
-interface userInfo {
+interface useGetInformation {
   phoneNumber: string,
   id: string,
   
   getUser: (tag: string , id: string ) => void;
 }
-export const useGetInformation = create<userInfo>(set => ({
+export const useGetInformation = create<useGetInformation>(set => ({
   phoneNumber: "",
   id:"",
   getUser: (number: string , id: string ): void => set({phoneNumber:number,id:id}),
+}));
+interface UseTextSMS {
+  textSMS: {value?:string, error?: boolean, message?: string}
+  setTextSMS: ({value, error, message}:{value?: string , error?: boolean, message?: string} ) => void;
+}
+export const useTextSMS = create<UseTextSMS>(set => ({
+  textSMS: {value:"", error: false, message: ""}
+  setTextSMS: ({value , error, message}) => set({textSMS:{
+        ...(value && {value}),
+        ...(error && {error}),
+        ...(message && {message}),
+      }
+    })
 }));
